requirements-worker.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2023.3
alpaca-py>=0.31.0
anthropic>=0.40.0
psycopg2-binary>=2.9.0
